requirements.txt
numpy==1.19.2
joblib
scipy
scikit-learn
pandas
streamlit
st-annotated-text
matplotlib
pybase64

